requirements.txt
pyqt5
pyqtwebengine
flask==1.1.2
pillow
